requirements.txt
appium-python-client==5.0.0
selenium>=4.10.0
pytest>=8.0.0
pytest-html>=4.0.0